+++ components/core/demos/design/brand-font.js
@@ -8,7 +8,11 @@ const StyledText = ({ weight, ...rest }) => {
 	const { PACKS, TYPE } = useBrand();
 	return (
 		<p
-			css={{ ...PACKS.typeScale.brandFont[4], fontWeight: TYPE.bodyFont[weight].fontWeight }}
+			css={{
+				...PACKS.typeScale.brandFont[4],
+				fontWeight: TYPE.bodyFont[weight].fontWeight,
+				overflowWrap: 'anywhere',
+			}}
 			{...rest}
 		/>
 	);
@@ -48,10 +52,14 @@ const Demo = ({ context, showCode, showDemo }) => {
 					<Fragment key={i}>
 						<p>{font.name}</p>
 						<StyledText weight={font.weight}>
+							abcdefghijklmnopqrstuvwxyz
+							<br />
 							ABCDEFGHIJKLMNOPQRSTUVWXYZ
 							<br />
-							abcdefghijklmnopqrstuvwxyz
-							<br />0 1 2 3 4 5 6 7 8 9
+							1234567890
+							<br />
+							!@#$¢%&amp;*©®™£
+							<br />ﬁ ﬂ ﬀ
 						</StyledText>
 					</Fragment>
 				))}