+++ components/selector/examples/50-disabled-state.js
@@ -8,6 +8,20 @@ function Example({ brand }) {
 		<GEL brand={brand}>
 			<h2>Global disabled</h2>
 
+			<h3>Button</h3>
+			<Selector type="button" name="default-button" disabled>
+				<Option value="1">Option 1</Option>
+				<Option value="2">Option 2</Option>
+				<Option value="3">Option 3</Option>
+			</Selector>
+
+			<h3>Submit</h3>
+			<Selector type="submit" name="default-submit" disabled>
+				<Option value="1">Option 1</Option>
+				<Option value="2">Option 2</Option>
+				<Option value="3">Option 3</Option>
+			</Selector>
+
 			<h3>Radio</h3>
 			<Selector type="radio" name="default-radio" disabled>
 				<Option value="1">Option 1</Option>
@@ -25,6 +39,22 @@ function Example({ brand }) {
 			<hr />
 
 			<h2>Disabled specific options</h2>
+
+			<h3>Button</h3>
+			<Selector type="button" name="default-button-specific" defaultValue={['2']}>
+				<Option value="1">Option 1</Option>
+				<Option value="2" disabled>
+					Option 2
+				</Option>
+			</Selector>
+
+			<h3>Submit</h3>
+			<Selector type="submit" name="default-submit-specific" defaultValue={['2']}>
+				<Option value="1">Option 1</Option>
+				<Option value="2" disabled>
+					Option 2
+				</Option>
+			</Selector>
 
 			<h3>Radio</h3>
 			<Selector type="radio" name="default-radio-specific" defaultValue={['2']}>
@@ -48,6 +78,22 @@ function Example({ brand }) {
 			<h2>Disabled fieldset</h2>
 
 			<fieldset disabled>
+				<h3>Button</h3>
+				<Selector type="button" name="default-button-fieldset" defaultValue={['2']}>
+					<Option value="1">Option 1</Option>
+					<Option value="2" disabled>
+						Option 2
+					</Option>
+				</Selector>
+
+				<h3>Submit</h3>
+				<Selector type="submit" name="default-submit-fieldset" defaultValue={['2']}>
+					<Option value="1">Option 1</Option>
+					<Option value="2" disabled>
+						Option 2
+					</Option>
+				</Selector>
+
 				<h3>Radio</h3>
 				<Selector type="radio" name="default-radio-fieldset" defaultValue={['2']}>
 					<Option value="1">Option 1</Option>