+++ website/dynamic-blocks/intro-section.js
@@ -13,9 +13,10 @@ import merge from 'lodash.merge';
 import { Body } from '../src/components/body';
 import { Icon } from '../../components/icon/src/Icon';
 import { Section } from '../src/components/section';
+import { ExternalLinkIcon } from '../src/components/external-link-icon';
 
 const ArrowDownRightIcon = (props) => (
-	<Icon assistiveText="Link arrow" {...props}>
+	<Icon aria-hidden="true" {...props}>
 		<path
 			fill="currentColor"
 			fillRule="evenodd"
@@ -79,10 +80,11 @@ const TableOfContents = ({ content }) => {
 
 	return (
 		<Fragment>
-			<nav ref={introRef}>
+			<nav ref={introRef} aria-labelledby="toc-heading">
 				<Heading
 					tag="h2"
 					size={9}
+					id="toc-heading"
 					overrides={{
 						Heading: {
 							styles: (styles) => ({
@@ -141,15 +143,15 @@ const TableOfContents = ({ content }) => {
 
 const PackageInfoTable = ({ item }) => {
 	if (!item) return null;
-	const { PACKS, SPACING } = useBrand();
+	const { PACKS, SPACING, COLORS } = useBrand();
 	const mq = useMediaQuery();
 	return (
 		<table
 			css={mq({
-				borderTop: 'solid 1px #2585ca',
+				borderTop: `1px solid ${COLORS.tints.info50}`,
 				borderCollapse: 'collapse',
-				background: '#f2f8fc',
-				color: '#2585ca',
+				backgroundColor: COLORS.tints.info5,
+				color: COLORS.info,
 				width: '100%',
 				textAlign: 'left',
 				...PACKS.typeScale.bodyFont[10],
@@ -161,7 +163,7 @@ const PackageInfoTable = ({ item }) => {
 						fontWeight: 500,
 					},
 					'> tr': {
-						borderBottom: 'solid 1px #2585CA',
+						borderBottom: `1px solid ${COLORS.tints.info50}`,
 						textAlign: 'left',
 						'> td, > th': { padding: SPACING(3) },
 					},
@@ -179,6 +181,7 @@ const PackageInfoTable = ({ item }) => {
 							target="_blank"
 						>
 							View Changes
+							<ExternalLinkIcon />
 						</a>
 					</td>
 				</tr>